DRY up the build with docker build-arg
Commit 3f3af200 added build-args for the runtime, so
it makes sense to use the same build arg for the builder image.
This allows to change the version in exactly one place: circle.yml

+++ script/build.sh
@@ -1,10 +1,19 @@
 #!/bin/bash
 set -e
 
+if [[ -z "${VERSION}" ]]; then
+  echo "You must export the VERSION environment variable to build" >&2
+  echo "See circle.yml for an example"
+  exit 1
+fi
+
 . script/functions
 
 # Build the authproxy.
-smitty docker build -t duoauthproxy-builder builder/
+smitty docker build \
+  --build-arg VERSION=${VERSION} \
+  -t duoauthproxy-builder \
+  builder/
 
 # Remove artifacts from previous runs.
 smitty rm -fr duoauthproxy.tgz || :
@@ -14,7 +23,6 @@ docker rm -f builder &> /dev/null || :
 smitty docker run --name builder duoauthproxy-builder bash -c "tar czf /tmp/duoauthproxy.tgz /opt/duoauthproxy"
 
 # Copy tarball into place and build runtime image.
-# NOTE: The build args are set by environment vars in circle.yml
 smitty docker cp builder:/tmp/duoauthproxy.tgz runtime/
 smitty docker build \
   --build-arg CI_BUILD_URL=${CIRCLE_BUILD_URL} \
